Setup trait split
