+++ src/WmiScripting/Concerns/HasHiddenAttributes.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PhpWinTools\WmiScripting\Concerns;
+
+trait HasHiddenAttributes
+{
+
+}
